Extract 'processed envelopes' queue classes to separate package
